% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateObject.R
\name{annotateObject}
\alias{annotateObject}
\alias{objectAnnotation}
\alias{setAnnotation}
\title{Add CollaboratorDB annotation}
\usage{
annotateObject(
  x,
  title,
  description,
  authors,
  species,
  genome,
  origin,
  terms = list()
)

objectAnnotation(x)

setAnnotation(x, annotation)
}
\arguments{
\item{x}{An \linkS4class{Annotated} object.}

\item{title}{String containing the object's title.}

\item{description}{String containing the description of the object.}

\item{authors}{Character vector containing the identity of the authors, in terms of their GitHub user names.}

\item{species}{An integer vector of NCBI taxonomy IDs for the species relevant to the data in \code{x}.}

\item{genome}{List describing the genomes involved in constructing \code{x}.
Each entry should be a list containing:
\itemize{
\item \code{id}, a string containing the genome build ID (e.g., hg38, NCBIm37)
\item \code{type}, a string specifying the type of ID (e.g., UCSC, Ensembl)
}
Check out \url{https://artifactdb.github.io/CollaboratorDB-schemas/array/v1.html#allOf_i0_genome} for details.}

\item{origin}{List describing the origin of \code{x}.
Each entry should be a list containing:
\itemize{
\item \code{source}, a string describing the source repository, e.g., PubMed, GEO, ArrayExpress.
\item \code{id}, a string containing an identifier within \code{source}.
}
Check out \url{https://artifactdb.github.io/CollaboratorDB-schemas/array/v1.html#allOf_i0_origin} for details.}

\item{terms}{List of ontology terms relevant to this resource.
Each entry should be a list containing:
\itemize{
\item \code{id}, a string containing an ID for an ontology term, e.g., \code{"EFO:0008913"}, \code{"DOID:13250"}, \code{"CL:0000097"}, \code{"UBERON:0005870"}.
\item \code{source}, a string containing the name of the ontology.
Currently, only Experimental Factor Ontology, Human Disease Ontology, Cell Ontology and UBERON are supported.
\item \code{version}, a string containing the version of the ontology.
This may be any tag on the corresponding GitHub repository for that ontology, e.g., \url{https://github.com/EBISPOT/efo/releases}.
}
Check out \url{https://artifactdb.github.io/CollaboratorDB-schemas/array/v1.html#allOf_i0_terms} for details.}
}
\value{
For \code{annotateObject}, \code{x} is returned with extra fields in its \code{\link{metadata}}.

For \code{objectAnnotation}, a list of CollaboratorDB-relevant metadata is returned.

For \code{setannotation}, \code{x} is returned after replacing the CollaboratorDB-relevant metadata with \code{annotation}.
}
\description{
Add CollaboratorDB's required annotation to a Bioconductor \linkS4class{Annotated} object.
}
\examples{
library(S4Vectors)
df <- DataFrame(X=1:10, Y=LETTERS[1:10], Z=factor(letters[1:10]))

df <- annotateObject(df,
    title="FOO",
    description="I am a data frame",
    authors="LTLA",
    species=9606,
    genome=list(list(id="hg38", source="UCSC")),
    origin=list(list(source="PubMed", id="123456789")),
    terms=list(list(id="EFO:0008913", source="Experimental Factor Ontology", version="v3.46.0"))
)

anno <- objectAnnotation(df)
str(anno)

anno$authors <- c(anno$authors, 
    list(list(name="Darth Vader", email="vader@empire.gov")))
df <- setAnnotation(df, anno)

}
\author{
Aaron Lun
}
