% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneDirectory.R
\name{cloneDirectory}
\alias{cloneDirectory}
\title{Clone a project to a directory}
\usage{
cloneDirectory(
  dir,
  project,
  version,
  link.only = TRUE,
  cache = TRUE,
  force.update = FALSE
)
}
\arguments{
\item{dir}{String containing a path to the destination directory.}

\item{project}{String containing the project name.}

\item{version}{String containing the project version.}

\item{link.only}{Logical scalar indicating whether to create placeholder links instead of copies of the non-metadata files.}

\item{cache}{Logical scalar indicating whether caching should be performed.
If \code{TRUE}, a default cache location is chosen.

Alternatively, a string specifying the path to a cache directory.

Alternatively, a \linkS4class{BiocFileCache} object.}

\item{force.update}{Logical scalar indicating whether cache entries should be forcibly updated.
Useful for fixing corrupted or incomplete files in the cache.#'}
}
\value{
A directory is created at \code{dir} with the contents of the specified project version.
\code{NULL} is invisibly returned.
}
\description{
Clone a specified version of an existing project into a directory,
typically for modifications prior to uploading a new version of the project via \code{\link{uploadDirectory}}.
}
\examples{
tmp <- tempfile()
cloneDirectory(tmp, "dssc-test_basic-2023", "2023-01-19")
list.files(tmp, recursive=TRUE)

}
\author{
Aaron Lun
}
