% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollaboratorDBArray.R
\docType{class}
\name{CollaboratorDBArray}
\alias{CollaboratorDBArray}
\alias{CollaboratorDBArray-class}
\alias{CollaboratorDBMatrix-class}
\alias{matrixClass,CollaboratorDBArray-method}
\alias{DelayedArray,CollaboratorDBArraySeed-method}
\alias{CollaboratorDBArraySeed-class}
\alias{loadArray}
\alias{CollaboratorDBArraySeed}
\title{The CollaboratorDBArray class}
\usage{
CollaboratorDBArraySeed(id)

CollaboratorDBArray(id, ...)

\S4method{DelayedArray}{CollaboratorDBArraySeed}(seed)
}
\arguments{
\item{id}{String containing the CollaboratorDB identifier, or a list containing the unpacked components of an identifier (from \code{\link{unpackID}}).
For \code{CollaboratorDBArray}, this can also be a CollaboratorDBArraySeed object.}

\item{...}{Further arguments to pass to the CollaboratorDBArraySeed constructor.}

\item{seed}{A CollaboratorDBArraySeed object.}
}
\value{
The \code{CollaboratorDBArraySeed} constructor returns a CollaboratorDBArraySeed object.

The \code{CollaboratorDBArray} constructor and the \code{\link{DelayedArray}} method return a CollaboratorDBArray object (or a CollaboratorDBMatrix, for two dimensions).
}
\description{
The CollaboratorDBArray is a \pkg{DelayedArray} wrapper around a on-disk array downloaded from the CollaboratorDB backend.
Specifically, it remembers the zircon ID that was used to download the array, which can be used to avoid a redundant write and upload in \code{\link{saveObject}} and \code{\link{uploadDirectory}}, respectively.
The actual heavy-lifting is done by forwarding all operations to the internal \pkg{DelayedArray}-compatible seed object,
which may be any of the usual classes, e.g., a HDF5ArraySeed, H5SparseMatrixSeed or any of its delayed operations.
}
\details{
If \code{id} uses the \code{latest} version alias, this is automatically resolved to a full version by the \code{CollaboratorDBArraySeed} constructor.
This ensures that the version is explicitly pinned in downstream applications and when saving to file.
}
\examples{
\dontshow{fun <- CollaboratorDB:::.configure_cache()}
id <- "test:my_first_sce/assay-1/matrix.h5@v1"
mat <- CollaboratorDBArray(id)
mat
\dontshow{fun()}

}
