% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchAllObjects.R
\name{fetchAllObjects}
\alias{fetchAllObjects}
\title{Fetch all objects for a project}
\usage{
fetchAllObjects(project, version = NULL, ...)
}
\arguments{
\item{project}{String containing the name of the project.}

\item{version}{String containing the name of the version.}

\item{...}{Further arguments to pass to \code{\link{fetchObject}}.}
}
\value{
If \code{version} is specified, a named list is returned containing all (non-child) R objects generated loading each resource in that version of the project.
The relative path to each resource is used as the name of the corresponding list element.

If \code{version=NULL}, a named list is returned containing one inner list per version.
Each inner list is named after its version and contains all (non-child) R objects for that version.
}
\description{
Fetch and load all objects for a project or one of its versions.
}
\details{
Only non-child objects are returned in each list.
Children are used to assemble their parent objects and are not reported here.

For redirections, a copy of the destination object is reported under the redirection's path.
This means that some resources may be reported multiple times - once for each redirection, and one for its \dQuote{true} path.
}
\examples{
# Specific version:
fetchAllObjects("dssc-test_basic-2023", version="2023-01-19")

# All available versions:
fetchAllObjects("dssc-test_basic-2023")

}
\author{
Aaron Lun
}
