% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listObjects.R
\name{listObjects}
\alias{listObjects}
\title{List objects for a project}
\usage{
listObjects(
  project,
  version = NULL,
  exclude.child = TRUE,
  ignore.redirection.target = TRUE
)
}
\arguments{
\item{project}{String containing the name of the project.}

\item{version}{String containing the name of the version.}

\item{exclude.child}{Logical scalar indicating whether to exclude child objects.}

\item{ignore.redirection.target}{Logical scalar indicating whether redirection targets should be ignored in the returned listing.}
}
\value{
If \code{version} is specified, a data frame is returned containing the path and ID for each object,
along with the metadata fields described in \code{\link{annotateObject}} (if available).

If \code{version=NULL}, a named list is returned containing one data frame per version.
Each data frame is named after its version and lists all objects for that version.
}
\description{
List available objects for a project or one of its versions.
}
\details{
Only non-child objects are guaranteed to contain the additional metadata fields described in \code{\link{annotateObject}}.
Child objects may or may not contain extra metadata, as their parents are assumed to take responsibility for describing them.

By default, redirection targets are ignored when reporting objects.
For example, if there is a redirection from path A to path B, the redirected path A is reported in the listing, and the redirection target B is ignored.
Setting \code{ignore.redirection.target=FALSE} will cause both A and B to be reported, so the same object will be present under different paths.
}
\examples{
# Specific version:
listObjects("dssc-test_basic-2023", version="2023-01-19")

# All available versions:
listObjects("dssc-test_basic-2023")

}
\author{
Aaron Lun
}
