% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchAllObjects.R
\name{fetchAllObjects}
\alias{fetchAllObjects}
\title{Fetch all objects for a project}
\usage{
fetchAllObjects(project, version = NULL, ..., ignore.redirection.target = TRUE)
}
\arguments{
\item{project}{String containing the name of the project.}

\item{version}{String containing the name of the version.}

\item{...}{Further arguments to pass to \code{\link{fetchObject}}.}

\item{ignore.redirection.target}{Logical scalar indicating whether redirection targets should be ignored in the returned listing.}
}
\value{
If \code{version} is specified, a named list is returned containing all (non-child) R objects generated loading each resource in that version of the project.
The relative path to each resource is used as the name of the corresponding list element.

If \code{version=NULL}, a named list is returned containing one inner list per version.
Each inner list is named after its version and contains all (non-child) R objects for that version.
}
\description{
Fetch and load all (non-child) objects for a project or one of its versions.
}
\details{
Only non-child objects are returned in each list.
Children are used to assemble their parent objects and are not reported here.

By default, redirection targets are ignored when reporting objects.
For example, if there is a redirection from path A to path B, the redirected path A is reported in the listing, and the redirection target B is ignored.
Setting \code{ignore.redirection.target=FALSE} will cause both A and B to be reported, so the same object will be present under different paths.
}
\examples{
# Specific version:
fetchAllObjects("dssc-test_basic-2023", version="2023-01-19")

# All available versions:
fetchAllObjects("dssc-test_basic-2023")

}
\author{
Aaron Lun
}
